Hide FAB on Log tab, move Delete button to toolbar menu
The FAB now animates when moving between tabs, though it no longer
appears on the Blocked Calls tab to match the Material spec, with the
action being moved to the toolbar menu.

"Avoid using floating action buttons for minor and destructive actions,
including the following:

* Archive or Trash"

Dialogs for deleting the blacklist and log are rewritten; the old
strings still remain, however. Machine-translated Spanish versions of
the new strings are added too, but I can't exactly vouch for their
quality.

ihavenoideawhatimdoing.jpg